Began Implementation of the AI for the dungeon monsters, and a bugfix
Created the base scripts for AI behavior trees, and implemented a handful of nodes for those trees and applied them to a base class for the monsters. Also fixed a bug that caused two versions of the basic ground to be added to the rooms list and caused all builders to occasionally try to pathfind to the same location.