Refined Order System and a bugfix
Edited the order system to no longer put orders in the high priority queue unless they are reachable, substantially decreasing the amount of time it takes for orders to get selected. Also fixed a bug that prevented orders from being assigned to cells that were not present at the start of the game.